refactor desktop workspace controls
